+++ easyPythonpi/__init__.py
@@ -1,1 +1,10 @@
 from easyPythonpi.easyPythonpi import *
+from easyPythonpi.methods.array import *
+from easyPythonpi.methods.basics import *
+from easyPythonpi.methods.matrix import *
+from easyPythonpi.methods.linkedlist import *
+from easyPythonpi.methods.sorting import *
+from easyPythonpi.methods.Graph import *
+from easyPythonpi.methods.search import *
+
+
