+++ easyPythonpi/methods/sorting.py
@@ -1,0 +1,9 @@
+
+def sort(list:'list'):    # To bubble sort and array or list
+    for i in range(len(list) - 1, 0, -1):
+        for j in range(i):
+            if list[j] > list[j + 1]:
+                temp = list[j]
+                list[j] = list[j + 1]
+                list[j + 1] = temp
+
